Fixed checksums
Fixed how the checksums are calculated and they now work.
The problem now is that when the host sends a SYN to the portforwarder, the portforwarder sends a RST packet right away.
Also, we need to get the portforwarders IP address some how, either as a parameter in the config file, or through the network interfaces.